+++ Src/App/Desktop/ScalesDesktop/Source/Shared/Models/WeightKneadingModel.cs
@@ -4,5 +4,5 @@ public class WeightKneadingModel
 {
     public int NetWeightG { get; set; }
     public DateTime ProductDate { get; set; } = DateTime.Now;
-    public int KneadingCount { get; set; } = 1;
+    public ushort KneadingCount { get; set; } = 1;
 }
